SecureNewsHub is a comprehensive news application prioritizing user security and personalized content delivery. It features robust authentication using JWT and offers a seamless experience with options like 2FA and passwordless login.
SecureNewsHub is a cutting-edge news platform designed to provide users with a secure and personalized news-reading experience. The project encompasses a secure RESTful API with JWT authentication, a GraphQL API for efficient news data retrieval, and a dynamic frontend for a user-friendly interface.
Explore the future of news consumption with SecureNewsHub, where security meets personalized content delivery. Fork, contribute, and be part of shaping the next generation of news applications!